made ping functions protocol independent
